requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.9
attrs==25.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.2
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
dill==0.3.9
distlib==0.3.9
dj-notification-api==1.2.1
Django==5.1.7
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-filter==25.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-rest-swagger==2.2.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.17.0
flake8==7.1.2
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
identify==2.6.9
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
inflection==0.5.1
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.2
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pylint==3.3.6
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
service-identity==24.2.0
setuptools==76.0.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
zope.interface==7.2
